requirements.txt
altair==5.4.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.0
grpcio-status==1.62.3
gTTS==2.5.3
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.5.5
networkx==3.3
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.23.0
proto-plus==1.24.0
protobuf==4.25.4
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyobjc==10.3.1
pyobjc-core==10.3.1
pyobjc-framework-Accessibility==10.3.1
pyobjc-framework-Accounts==10.3.1
pyobjc-framework-AddressBook==10.3.1
pyobjc-framework-AdServices==10.3.1
pyobjc-framework-AdSupport==10.3.1
pyobjc-framework-AppleScriptKit==10.3.1
pyobjc-framework-AppleScriptObjC==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-AppTrackingTransparency==10.3.1
pyobjc-framework-AudioVideoBridging==10.3.1
pyobjc-framework-AuthenticationServices==10.3.1
pyobjc-framework-AutomaticAssessmentConfiguration==10.3.1
pyobjc-framework-Automator==10.3.1
pyobjc-framework-AVFoundation==10.3.1
pyobjc-framework-AVKit==10.3.1
pyobjc-framework-AVRouting==10.3.1
pyobjc-framework-BackgroundAssets==10.3.1
pyobjc-framework-BrowserEngineKit==10.3.1
pyobjc-framework-BusinessChat==10.3.1
pyobjc-framework-CalendarStore==10.3.1
pyobjc-framework-CallKit==10.3.1
pyobjc-framework-CFNetwork==10.3.1
pyobjc-framework-Cinematic==10.3.1
pyobjc-framework-ClassKit==10.3.1
pyobjc-framework-CloudKit==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Collaboration==10.3.1
pyobjc-framework-ColorSync==10.3.1
pyobjc-framework-Contacts==10.3.1
pyobjc-framework-ContactsUI==10.3.1
pyobjc-framework-CoreAudio==10.3.1
pyobjc-framework-CoreAudioKit==10.3.1
pyobjc-framework-CoreBluetooth==10.3.1
pyobjc-framework-CoreData==10.3.1
pyobjc-framework-CoreHaptics==10.3.1
pyobjc-framework-CoreLocation==10.3.1
pyobjc-framework-CoreMedia==10.3.1
pyobjc-framework-CoreMediaIO==10.3.1
pyobjc-framework-CoreMIDI==10.3.1
pyobjc-framework-CoreML==10.3.1
pyobjc-framework-CoreMotion==10.3.1
pyobjc-framework-CoreServices==10.3.1
pyobjc-framework-CoreSpotlight==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-CoreWLAN==10.3.1
pyobjc-framework-CryptoTokenKit==10.3.1
pyobjc-framework-DataDetection==10.3.1
pyobjc-framework-DeviceCheck==10.3.1
pyobjc-framework-DictionaryServices==10.3.1
pyobjc-framework-DiscRecording==10.3.1
pyobjc-framework-DiscRecordingUI==10.3.1
pyobjc-framework-DiskArbitration==10.3.1
pyobjc-framework-DVDPlayback==10.3.1
pyobjc-framework-EventKit==10.3.1
pyobjc-framework-ExceptionHandling==10.3.1
pyobjc-framework-ExecutionPolicy==10.3.1
pyobjc-framework-ExtensionKit==10.3.1
pyobjc-framework-ExternalAccessory==10.3.1
pyobjc-framework-FileProvider==10.3.1
pyobjc-framework-FileProviderUI==10.3.1
pyobjc-framework-FinderSync==10.3.1
pyobjc-framework-FSEvents==10.3.1
pyobjc-framework-GameCenter==10.3.1
pyobjc-framework-GameController==10.3.1
pyobjc-framework-GameKit==10.3.1
pyobjc-framework-GameplayKit==10.3.1
pyobjc-framework-HealthKit==10.3.1
pyobjc-framework-ImageCaptureCore==10.3.1
pyobjc-framework-InputMethodKit==10.3.1
pyobjc-framework-InstallerPlugins==10.3.1
pyobjc-framework-InstantMessage==10.3.1
pyobjc-framework-Intents==10.3.1
pyobjc-framework-IntentsUI==10.3.1
pyobjc-framework-IOBluetooth==10.3.1
pyobjc-framework-IOBluetoothUI==10.3.1
pyobjc-framework-IOSurface==10.3.1
pyobjc-framework-iTunesLibrary==10.3.1
pyobjc-framework-KernelManagement==10.3.1
pyobjc-framework-LatentSemanticMapping==10.3.1
pyobjc-framework-LaunchServices==10.3.1
pyobjc-framework-libdispatch==10.3.1
pyobjc-framework-libxpc==10.3.1
pyobjc-framework-LinkPresentation==10.3.1
pyobjc-framework-LocalAuthentication==10.3.1
pyobjc-framework-LocalAuthenticationEmbeddedUI==10.3.1
pyobjc-framework-MailKit==10.3.1
pyobjc-framework-MapKit==10.3.1
pyobjc-framework-MediaAccessibility==10.3.1
pyobjc-framework-MediaLibrary==10.3.1
pyobjc-framework-MediaPlayer==10.3.1
pyobjc-framework-MediaToolbox==10.3.1
pyobjc-framework-Metal==10.3.1
pyobjc-framework-MetalFX==10.3.1
pyobjc-framework-MetalKit==10.3.1
pyobjc-framework-MetalPerformanceShaders==10.3.1
pyobjc-framework-MetalPerformanceShadersGraph==10.3.1
pyobjc-framework-MetricKit==10.3.1
pyobjc-framework-MLCompute==10.3.1
pyobjc-framework-ModelIO==10.3.1
pyobjc-framework-MultipeerConnectivity==10.3.1
pyobjc-framework-NaturalLanguage==10.3.1
pyobjc-framework-NetFS==10.3.1
pyobjc-framework-Network==10.3.1
pyobjc-framework-NetworkExtension==10.3.1
pyobjc-framework-NotificationCenter==10.3.1
pyobjc-framework-OpenDirectory==10.3.1
pyobjc-framework-OSAKit==10.3.1
pyobjc-framework-OSLog==10.3.1
pyobjc-framework-PassKit==10.3.1
pyobjc-framework-PencilKit==10.3.1
pyobjc-framework-PHASE==10.3.1
pyobjc-framework-Photos==10.3.1
pyobjc-framework-PhotosUI==10.3.1
pyobjc-framework-PreferencePanes==10.3.1
pyobjc-framework-PushKit==10.3.1
pyobjc-framework-Quartz==10.3.1
pyobjc-framework-QuickLookThumbnailing==10.3.1
pyobjc-framework-ReplayKit==10.3.1
pyobjc-framework-SafariServices==10.3.1
pyobjc-framework-SafetyKit==10.3.1
pyobjc-framework-SceneKit==10.3.1
pyobjc-framework-ScreenCaptureKit==10.3.1
pyobjc-framework-ScreenSaver==10.3.1
pyobjc-framework-ScreenTime==10.3.1
pyobjc-framework-ScriptingBridge==10.3.1
pyobjc-framework-SearchKit==10.3.1
pyobjc-framework-Security==10.3.1
pyobjc-framework-SecurityFoundation==10.3.1
pyobjc-framework-SecurityInterface==10.3.1
pyobjc-framework-SensitiveContentAnalysis==10.3.1
pyobjc-framework-ServiceManagement==10.3.1
pyobjc-framework-SharedWithYou==10.3.1
pyobjc-framework-SharedWithYouCore==10.3.1
pyobjc-framework-ShazamKit==10.3.1
pyobjc-framework-Social==10.3.1
pyobjc-framework-SoundAnalysis==10.3.1
pyobjc-framework-Speech==10.3.1
pyobjc-framework-SpriteKit==10.3.1
pyobjc-framework-StoreKit==10.3.1
pyobjc-framework-Symbols==10.3.1
pyobjc-framework-SyncServices==10.3.1
pyobjc-framework-SystemConfiguration==10.3.1
pyobjc-framework-SystemExtensions==10.3.1
pyobjc-framework-ThreadNetwork==10.3.1
pyobjc-framework-UniformTypeIdentifiers==10.3.1
pyobjc-framework-UserNotifications==10.3.1
pyobjc-framework-UserNotificationsUI==10.3.1
pyobjc-framework-VideoSubscriberAccount==10.3.1
pyobjc-framework-VideoToolbox==10.3.1
pyobjc-framework-Virtualization==10.3.1
pyobjc-framework-Vision==10.3.1
pyobjc-framework-WebKit==10.3.1
pyparsing==3.1.4
python-box==7.2.0
python-dateutil==2.9.0.post0
pyttsx3==2.91
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
streamlit==1.38.0
streamlit-scrollable-textbox==0.0.3
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
torch==2.4.0
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
